Move migration of jobs to GitLabPushTrigger to remove deprecated getters
